Olemassa olevan usernimen tsekkaus

+++ pages/accCreation/accCreation.ts
@@ -1,5 +1,5 @@
 import { doGraphQLFetch } from "../../src/graphql/fetch"
-import { allUsers, createUser } from "../../src/graphql/queries"
+import { allUsers, createUser, userByUsername } from "../../src/graphql/queries"
 import LoginMessageResponse from "../../src/interfaces/LoginMessageResponse";
 import { User, UserIdWithToken } from "../../src/interfaces/User"
 
@@ -20,14 +20,25 @@ registerButton.addEventListener("click", async (e) => {
   const nameInput = username.value;
   const passInput = password.value;
   try {
-    const registerData = (await doGraphQLFetch(apiURL, createUser, {
-        user: {
-          username: nameInput,
-          password: passInput
-        }
-    })) as LoginMessageResponse;
-    localStorage.setItem("token", registerData.login?.token!);
-    window.location.href = '../../pages/dogPark/index.html';
+    const existingUser = (await doGraphQLFetch(apiURL, userByUsername, {
+      username: nameInput
+    }))
+
+    if (!existingUser.userByUsername) {
+      const registerData = (await doGraphQLFetch(apiURL, createUser, {
+          user: {
+            username: nameInput,
+            password: passInput
+          }
+      })) as LoginMessageResponse;
+      localStorage.setItem("token", registerData.login?.token!);
+      window.location.href = '../../pages/dogPark/index.html';
+    } else {
+      const regInfo = document.querySelector("#regInfo") as HTMLElement;
+      if (regInfo) {
+        regInfo.textContent = `Käyttäjätunnus virheellinen. Valitse toinen.`;
+      }
+    }
   } catch (error) {
     console.log(error);
   }
